+++ flask_cuttlepool.py
@@ -29,12 +29,12 @@ class FlaskCuttlePool(CuttlePool):
     :param func connect: The ``connect`` function of the chosen sql driver.
     :param int capacity: Max number of connections in pool. Defaults to ``5``.
     :param int timeout: Time in seconds to wait for connection. Defaults to
-                        ``None``.
+        ``None``.
     :param int overflow: The number of extra connections that can be made if
-                         the pool is exhausted. Defaults to ``1``.
+        the pool is exhausted. Defaults to ``1``.
     :param Flask app: A Flask ``app`` object. Defaults to ``None``.
     :param \**kwargs: Connection arguments for the underlying database
-                      connector.
+        connector.
     """
 
     def __init__(self, connect, capacity=5, overflow=1,
@@ -46,30 +46,32 @@ class FlaskCuttlePool(CuttlePool):
         if app is not None:
             self.init_app(app)
 
-    def init_app(self, app, **kwargs):
+    def init_app(self, app):
         """
         Configures the connection pool and attaches a teardown handler to the
-        ``app`` object.
+        ``app`` object. All configuration options on ``app.config`` of the form
+        ``CUTTLEPOOL_<KEY>`` will be used as connection arguments for the
+        underlying driver. ``<KEY>`` will be converted to lowercase such that
+        ``app.config['CUTTLEPOOL_<KEY>'] = <value>`` will be passed to the
+        connection driver as ``<key>=<value>``.
 
         :param Flask app: A Flask ``app`` object.
-        :param \**kwargs: Connection arguments for the underlying database
-                          connector. These will overwrite any key value pairs
-                          with the same key that were passed to
-                          ``FlaskCuttlePool`` on instantiation. This is where
-                          any arguments from ``app.config`` should be passed to
-                          to ``FlaskCuttlePool``.
 
         :Example:
 
         # pool will connect to rons_house.
         pool = FlaskCuttlePool(sqlite3.connect, database='rons_house')
         app = Flask(__name__)
-        app.config['SQLITE3_DB'] = 'steakhouse'
+        app.config['CUTTLEPOOL_DATABASE'] = 'steakhouse'
         # pool will connect to steakhouse instead.
-        pool.init_app(app, database=app.config['SQLITE3_DB'])
+        pool.init_app(app)
         """
+        prefix = 'CUTTLEPOOL_'
         self._app_set = True
-        self._connection_arguments.update(**kwargs)
+        self._connection_arguments.update(
+            **{k[len(prefix):].lower(): v
+               for k, v in app.config.items()
+               if k.startswith(prefix)})
 
         # Use the newstyle teardown_appcontext if it's available,
         # otherwise fall back to the request context.